+++ src/server/replicated_archetypes.rs
@@ -16,7 +16,8 @@ use crate::core::{replication_fns::SerdeFnsId, Replication};
 /// - Register 'serde' and 'remove' functions inside [`ReplicationFns`](crate::core::replication_fns::ReplicationFns).
 /// - Update this struct for all newly added archetypes in
 /// [`ServerSet::UpdateArchetypes`](super::ServerSet::UpdateArchetypes) using the registered function IDs.
-/// - Update [`RemovalBuffer`](super::removal_buffer::RemovalBuffer) when your special components get removed.
+/// - Update [`RemovalBuffer`](super::removal_buffer::RemovalBuffer) in
+/// [`ServerSet::BufferRemovals`](super::ServerSet::BufferRemovals) when the rule components should be removed.
 #[derive(Resource)]
 pub struct ReplicatedArchetypes {
     archetypes: Vec<ReplicatedArchetype>,